# backend/docs/swagger.json
{
    "produces": [
        "application/json"
    ],
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "Backend API for the Tealive \u0026 Baskbear Twin App",
        "title": "Loob Unified API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
        "/catalog/categories": {
            "get": {
                "description": "Lists localized menu categories without product payloads",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "catalog"
                ],
                "summary": "List localized menu categories",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Store ID",
                        "name": "store_id",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Brand ID",
                        "name": "brand_id",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "default": "MY",
                        "description": "Country Code (e.g., MY, TH)",
                        "name": "X-Country-Code",
                        "in": "header"
                    },
                    {
                        "type": "string",
                        "default": "en-US",
                        "description": "Language (e.g., en-US, ms-MY)",
                        "name": "Accept-Language",
                        "in": "header"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/catalog/categories/{category_id}/items": {
            "get": {
                "description": "Lists localized menu items for a single category",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "catalog"
                ],
                "summary": "Get localized category items",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Category ID",
                        "name": "category_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Store ID",
                        "name": "store_id",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Brand ID",
                        "name": "brand_id",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "default": "MY",
                        "description": "Country Code (e.g., MY, TH)",
                        "name": "X-Country-Code",
                        "in": "header"
                    },
                    {
                        "type": "string",
                        "default": "en-US",
                        "description": "Language (e.g., en-US, ms-MY)",
                        "name": "Accept-Language",
                        "in": "header"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/catalog/brands": {
            "get": {
                "description": "List all available brands",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "catalog"
                ],
                "summary": "List brands",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/Brand"
                            }
                        }
                    }
                }
            }
        },
        "/catalog/stores": {
            "get": {
                "description": "List stores by country",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "catalog"
                ],
                "summary": "List stores",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Country Code Filter",
                        "name": "country_id",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/Store"
                            }
                        }
                    }
                }
            }
        },
        "/orders/checkout": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Pushes the order intent to the SQS queue for asynchronous processing",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "orders"
                ],
                "summary": "Process an order",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Country Code (e.g., MY, TH)",
                        "name": "X-Country-Code",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "202": {
                        "description": "Accepted",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    },
    "definitions": {
        "Brand": {
            "type": "object",
            "properties": {
                "id": { "type": "integer" },
                "slug": { "type": "string" },
                "name": { "type": "string" },
                "primary_color": { "type": "string" },
                "accent_color": { "type": "string" }
            }
        },
        "Store": {
            "type": "object",
            "properties": {
                "id": { "type": "integer" },
                "brand_id": { "type": "integer" },
                "country_id": { "type": "string" },
                "zone_id": { "type": "string" },
                "store_code": { "type": "string" },
                "name": { "type": "string" },
                "latitude": { "type": "number" },
                "longitude": { "type": "number" },
                "address": { "type": "string" },
                "is_active": { "type": "boolean" }
            }
        }
    }
}